// src/common/enums/story-era.enum.ts
export enum StoryEra {
  BCE = 'BCE',
  CE = 'CE',
  MODERN = 'MODERN',
  FIRST_AGE = 'FIRST_AGE',
  SECOND_AGE = 'SECOND_AGE',
  THIRD_AGE = 'THIRD_AGE',
  FOURTH_AGE = 'FOURTH_AGE',
  FOURTH_AGE_OF_MIDDLE_EARTH = 'FOURTH_AGE_OF_MIDDLE_EARTH',
  UNSPECIFIED = 'UNSPECIFIED',
}

export const STORY_ERA_VALUES: StoryEra[] = Object.values(StoryEra);