// src/common/enums/story-civilization.enum.ts
export enum StoryCivilization {
  ARABIC = 'ARABIC',
  EGYPTIAN = 'EGYPTIAN',
  ANCIENT_EGYPTIAN = 'ANCIENT_EGYPTIAN',
  GREEK = 'GREEK',
  ROMAN = 'ROMAN',
  CUSTOM = 'CUSTOM',
  UNSPECIFIED = 'UNSPECIFIED',
}

export const STORY_CIVILIZATION_VALUES: StoryCivilization[] =
  Object.values(StoryCivilization);